--- src/tw-styled/types/Theming.ts
@@ -1,3 +1,0 @@
-export type Themes = "light" | "dark";
-
-export type BtnVariants = "none" | "primary" | "warning" | "ghost";
