+++ src/tw-styled/types/index.ts
@@ -1,3 +1,1 @@
-export * from "./Elements";
-export * from "./Theming";
 export * from "../style-to-class-resolver/types";
