+++ src/ui/components/box/Box.types.ts
@@ -1,4 +1,5 @@
-import { BaseStyles, WrapperElements } from "@/tw-styled/types";
+import { BaseStyles } from "@/tw-styled/types";
+import { WrapperElements } from "@/ui/types";
 
 export type BoxProps = {
   children?: React.ReactNode;
